pr-changes

+++ src/controllers/jobCoordinator.ts
@@ -10,6 +10,8 @@ import {
   Body,
   Put,
   UseGuards,
+  HttpException,
+  HttpStatus,
 } from "@nestjs/common";
 import { AuthGuard } from "@nestjs/passport";
 import { ApiBearerAuth } from "@nestjs/swagger";
@@ -17,12 +19,16 @@ import { JOB_COORDINATOR_SERVICE } from "src/constants";
 import { JobIdParamDto } from "src/dtos/job";
 import JobCoordinatorService from "src/services/JobCoordinatorService";
 import { JobCoordinator } from "src/entities/JobCoordinator";
-import { JobCoordinatorIdParamDto, createJobCoordinatorDto, updateJobCoordinatorDto } from "src/dtos/jobCoordinator";
+import { CreateJobCoordinatorsDto, JobCoordinatorIdParamDto,  UpdateJobCoordinatorDto } from "src/dtos/jobCoordinator";
+import { TransactionInterceptor } from "src/interceptor/TransactionInterceptor";
+import { TransactionParam } from "src/decorators/TransactionParam";
+import { Transaction } from "sequelize";
+import { UpdateOrFind } from "src/utils/utils";
 
 @Controller("/job/:jobId/coordinators")
 @ApiBearerAuth("jwt")
 @UseGuards(AuthGuard("jwt"))
-export class PenaltyController {
+export class JobCoordinatorController {
   constructor(@Inject(JOB_COORDINATOR_SERVICE) private jobCoordinatorService: JobCoordinatorService) {}
 
   @Get()
@@ -34,27 +40,66 @@ export class PenaltyController {
 
   @Post()
   @UseInterceptors(ClassSerializerInterceptor)
-  async createJobCoordinator(@Param() param: JobIdParamDto, @Body() body: createJobCoordinatorDto) {
-    const newJobCoordinator = await this.jobCoordinatorService.createJobCoordinator(
-      new JobCoordinator({
-        jobId: param.jobId,
-        tpcMemberId: body.tpcMemberId,
-        role: body.role
-      })
-    );
-    return { jobCoordinator: newJobCoordinator };
+  @UseInterceptors(TransactionInterceptor)
+  async createJobCoordinators(@Param() param: JobIdParamDto, @Body() body: CreateJobCoordinatorsDto, @TransactionParam() transaction: Transaction) {
+    const promises = [];
+    for (const jobCoordinator of body.jobCoordinators) {
+      promises.push(
+        new Promise(async (resolve, reject) => {
+          try {
+            const newJobCoordinator = await this.jobCoordinatorService.createOrGetJobCoordinator(
+              new JobCoordinator({
+                jobId: param.jobId,
+                tpcMemberId: jobCoordinator.tpcMemberId,
+                role: jobCoordinator.role
+              }),
+              transaction
+            );
+            resolve(newJobCoordinator);
+          } catch (err) {
+            reject(err);
+          }
+        })
+      );
+    }
+    const jobCoordinators = await Promise.all(promises);
+    return { jobCoordinators: jobCoordinators };
   }
 
   @Put("/:jobCoordinatorId")
   @UseInterceptors(ClassSerializerInterceptor)
-  async updateJobCoordinator(@Param() param: JobIdParamDto & JobCoordinatorIdParamDto, @Body() body: updateJobCoordinatorDto) {
-    const newJobCoordinator = await this.jobCoordinatorService.updateJobCoordinator(param.jobCoordinatorId, body);
+  @UseInterceptors(TransactionInterceptor)
+  async updateJobCoordinator(
+    @Param() param: JobIdParamDto & JobCoordinatorIdParamDto, 
+    @Body() body: UpdateJobCoordinatorDto, 
+    @TransactionParam() transaction: Transaction) {
+    const [jobCoordinator] = await this.jobCoordinatorService.getJobCoordinators({
+      id: param.jobCoordinatorId,
+    });
+    if (!jobCoordinator) {
+      throw new HttpException(`jobCoordinator with jobCoordinatorId: ${param.jobCoordinatorId} not found`, HttpStatus.NOT_FOUND);
+    }
+    
+    const newJobCoordinator = await UpdateOrFind(
+      param.jobCoordinatorId,
+      body,
+      this.jobCoordinatorService,
+      "updateJobCoordinator",
+      "getJobCoordinators",
+      transaction
+    );
     return { jobCoordinator: newJobCoordinator };
   }
 
   @Delete("/:jobCoordinatorId")
   @UseInterceptors(ClassSerializerInterceptor)
   async deleteJobCoordinator(@Param() param: JobIdParamDto & JobCoordinatorIdParamDto) {
+    const [jobCoordinator] = await this.jobCoordinatorService.getJobCoordinators({
+      id: param.jobCoordinatorId,
+    });
+    if (!jobCoordinator) {
+      throw new HttpException(`jobCoordinator with jobCoordinatorId: ${param.jobCoordinatorId} not found`, HttpStatus.NOT_FOUND);
+    }
     const deleted = await this.jobCoordinatorService.deleteJobCoordinator(param.jobCoordinatorId);
     return { deleted: deleted };
   }
