+++ src/controllers/recruiter.ts
@@ -10,12 +10,14 @@ import {
   Param,
   Put,
   Delete,
+  HttpException,
+  HttpStatus,
 } from "@nestjs/common";
 import { RECRUITER_SERVICE, USER_SERVICE } from "src/constants";
 import RecruiterService from "src/services/RecruiterService";
 import UserService from "src/services/UserService";
 import { Recruiter } from "src/entities/Recruiter";
-import { AddRecruiterDto, RecruiterIdParamDto, UpdateRecruiterDto } from "../dtos/recruiter";
+import { AddRecruitersDto, RecruiterIdParamDto, UpdateRecruiterDto } from "../dtos/recruiter";
 import { ApiBearerAuth } from "@nestjs/swagger";
 import { AuthGuard } from "@nestjs/passport";
 import { TransactionInterceptor } from "src/interceptor/TransactionInterceptor";
@@ -24,6 +26,7 @@ import { Transaction } from "sequelize";
 import { User } from "src/entities/User";
 import { Role } from "src/db/enums";
 import { CompanyIdParamDto } from "src/dtos/company";
+import { UpdateOrFind } from "src/utils/utils";
 
 @Controller("/:companyId/recruiters")
 @ApiBearerAuth("jwt")
@@ -40,6 +43,9 @@ export class RecruiterController {
     const recruiters = await this.recruiterService.getRecruiters({
       companyId: param.companyId,
     });
+    for (const recruiter of recruiters) {
+      recruiter.user = await this.userService.getUserById(recruiter.userId);
+    }
     return { recruiters: recruiters };
   }
 
@@ -48,36 +54,91 @@ export class RecruiterController {
   @UseInterceptors(ClassSerializerInterceptor)
   async addRecruiter(
     @Param() param: CompanyIdParamDto,
-    @Body() body: AddRecruiterDto,
+    @Body() body: AddRecruitersDto,
     @TransactionParam() transaction: Transaction
   ) {
-    const user = await this.userService.getOrCreateUser(
-      new User({ name: body.name, email: body.email, contact: body.contact, role: Role.RECRUITER }),
-      transaction
-    );
+    const promises = [];
+    for (const recruiter of body.recruiters) {
+      promises.push(
+        new Promise(async (resolve, reject) => {
+          try {
+            const user = await this.userService.getOrCreateUser(
+              new User({
+                name: recruiter.name,
+                email: recruiter.email,
+                contact: recruiter.contact,
+                role: Role.RECRUITER,
+              }),
+              transaction
+            );
 
-    const newRecruiter = await this.recruiterService.createRecruiter(
-      new Recruiter({
-        userId: user.id,
-        companyId: param.companyId,
-      }),
-      transaction
-    );
-    newRecruiter.user = user;
+            const newRecruiter = await this.recruiterService.getOrCreateRecruiter(
+              new Recruiter({
+                userId: user.id,
+                companyId: param.companyId,
+              }),
+              transaction
+            );
+            newRecruiter.user = user;
 
-    return { recruiter: newRecruiter };
+            resolve(newRecruiter);
+          } catch (err) {
+            reject(err);
+          }
+        })
+      );
+    }
+
+    const Recruiters = await Promise.all(promises);
+    return { Recruiters: Recruiters };
+  }
+
+  querybuilder(params) {
+    let Recruiter = {};
+    let User = {};
+    if (params.companyId) {
+      Recruiter["companyId"] = params.companyId;
+    }
+    if (params.name) {
+      User["name"] = params.name;
+    }
+    if (params.email) {
+      User["email"] = params.email;
+    }
+    if (params.contact) {
+      User["contact"] = params.contact;
+    }
+    if (params.role) {
+      User["role"] = params.role;
+    }
+
+    return { Recruiter, User };
   }
 
   @Put("/:recruiterId")
   @UseInterceptors(ClassSerializerInterceptor)
   @UseInterceptors(TransactionInterceptor)
-  async updateStudent(
+  async updateRecruiter(
     @Param() param: CompanyIdParamDto & RecruiterIdParamDto,
     @Body() body: UpdateRecruiterDto,
     @TransactionParam() transaction: Transaction
   ) {
-    const newRecruiter = await this.recruiterService.updateRecruiter(param.recruiterId, body, transaction);
-    const newUser = await this.userService.updateUser(newRecruiter.userId, body, transaction);
+    const [recruiter] = await this.recruiterService.getRecruiters({
+      id: param.recruiterId,
+    });
+    if (!recruiter) {
+      throw new HttpException(`recruiter with recruiterId: ${param.recruiterId} not found`, HttpStatus.NOT_FOUND);
+    }
+    const { Recruiter, User } = this.querybuilder(body);
+    const newRecruiter = await UpdateOrFind(
+      param.recruiterId,
+      Recruiter,
+      this.recruiterService,
+      "updateRecruiter",
+      "getRecruiters",
+      transaction
+    );
+    const newUser = await UpdateOrFind(newRecruiter.userId, User, this.userService, "updateUser", "getUserById",transaction);
     newRecruiter.user = newUser;
     return { recruiter: newRecruiter };
   }
@@ -85,12 +146,19 @@ export class RecruiterController {
   @Delete("/:recruiterId")
   @UseInterceptors(ClassSerializerInterceptor)
   @UseInterceptors(TransactionInterceptor)
-  async deletePenalty(
+  async deleteRecruiter(
     @Param() param: CompanyIdParamDto & RecruiterIdParamDto,
     @TransactionParam() transaction: Transaction
   ) {
-    const userId = await this.recruiterService.deleteRecruiter(param.recruiterId, transaction);
+    const [recruiter] = await this.recruiterService.getRecruiters({
+      id: param.recruiterId,
+    });
+    if (!recruiter) {
+      throw new HttpException(`recruiter with recruiterId: ${param.recruiterId} not found`, HttpStatus.NOT_FOUND);
+    }
+    const userId = recruiter.userId;
+    const recruiterDeleted = await this.recruiterService.deleteRecruiter(param.recruiterId, transaction);
     const userDeleted = await this.userService.deleteUser(userId, transaction);
-    return { deleted: userDeleted };
+    return { deleted: userDeleted && recruiterDeleted };
   }
 }
